requirements-stdborg.txt
aiofiles
aiohttp
aria2p
beautifulsoup4
coffeehouse
cowpy
cfscrape
cryptg
emoji
googletrans
google_images_download
google-api-python-client
google_auth_oauthlib
gsearch>=1.6.0
gtts
hachoir
Pillow
psycopg2
pySmartDL
python-magic
qrcode
regex
requests
PyGithub
speedtest-cli
sqlalchemy
telegraph
urbandict
wikipedia
youtube-dl
lxml
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
oauth2client
httplib2
humanize
patool
spotify_token
fast.com
names
random_cli
pytuneteller
dadjokes-cli
jotquote
csvfaker
gitpython
heroku3
